latest fixes for telosb and nrf + configure serial forwarder on telosb

+++ raspi/scripts/nrf52/serialdump.sh
@@ -2,15 +2,12 @@
 
 log_path=$1
 tty_path=`ls /dev/serial/by-id/usb-SEGGER_J-Link_*`
-killall -9 picocom
-killall -9 screen
+if pgrep picocom; then killall -9 picocom; fi; 
+if pgrep screen; then killall -9 screen; fi; 
+if pgrep contiki-timestamp; then killall -9 contiki-timestamp; fi;
 screen -wipe
 screen -dmS nrf52screen bash
-screen -S nrf52screen -X stuff "picocom --noreset -fh -b 115200 --imap lfcrlf $tty_path | ~/scripts/contiki-timestamp > $log_path\n"
-#screen -S nrf52screen -X stuff "~/scripts/contiki-serialdump -b230400 $tty_path | ~/scripts/contiki-timestamp > $log_path\n"
-#nohup ~/scripts/contiki-serialdump -b230400 $tty_path | ~/scripts/contiki-timestamp > $log_path & > /dev/null 2> /dev/null
-#stty 230400 crtscts cs8
-#nohup sh -c "cat $tty_path | ~/scripts/contiki-timestamp > $log_path & > /dev/null 2> /dev/null" 
+screen -S nrf52screen -X stuff "picocom -fh -b 115200 --imap lfcrlf $tty_path | ~/scripts/contiki-timestamp > $log_path\n"
 sleep 1
 ps | grep "$! "
 exit $?
